checkpoint, committing then abandoning.
This was an attempt to get all fancy by identifying which macros
got expanded to what. But it has gotten messy and complicated,
and will require huge amounts of testing for corner cases, and
it's just not worth it. In the end, a human can diff two lines
better than what we can do.