Allow downloading builds from any Koji instance
This patch adds changes to rpmgrill-fetch-build to download rpms from
any Koji instance.

The script can be run with two additional parameters:

    --server-url which is the corresponding XMLRPC endpoint and
      defaults to: http://koji.fedoraproject.org/kojihub

    --top-url which corresponds to the HTTP server from which we download
      builds and build logs.

The _fetch_cached_rpms method has been deprecated. We could not figure
out a way to intuitively use it in combination with the --download-only
parameter. Furthermore it was not covered by tests, which makes
refactoring harder.

Bug: 1202634
Change-Id: Iec52c9c2c51e16d8d93726a870fd83010f2c5b60