tests/virus-scan: fixes fragile hard-coded test data
ClamAv scan would output the files in random order and the test
data had assumed the order in which the results appear. This change
fixes that by enabling the test data to be a regex. The caveat is
that this change assumes that the same file 'myfile1' or 'myfile2'
doesn't appear twice in ClamAv's output.

Change-Id: I9c72a16c4d9997bd619e4816c15054e4afa139b3