setxid: replace hard-coded file path with HOME
setxid breaks when RPMGRILL_SETXID_WHITELIST_DIR is not set and the
user running rpmgrill is not 'rpmgrill' since the path it searches for
the whitelist file was hard-coded to '/home/rpmgrill'. This patch fixes
that by replacing that with users home directory or whatever HOME
environment variable points to.

Change-Id: Idf085a652b4a59bfe4e162d2ac79a3b8b5dc3954