setxid: replace hard-coded file path with HOME
setxid prints a warning when RPMGRILL_SETXID_WHITELIST_DIR is not set
and the user running rpmgrill is not 'rpmgrill' as the path it searches
for the whitelist file was hard-coded to '/home/rpmgrill'.
This patch fixes the hard-coded path by replacing it with
~/.rpmgrill/plugins/setxid where `~` is the path pointed by $HOME
environment variable.

Additionally, the patch also prints a workaround for suppressing the
warning.

Change-Id: Idf085a652b4a59bfe4e162d2ac79a3b8b5dc3954