docker: refactor images to base and dev
This patch splits the existing  docker images into base and dev.
It moves the steps common to all images to 'base'.  So the base has
minimal set of tools like git, setup, vim, initscripts.

The 'dev' image builds on top of base and adds the dependencies.

The unused 'CI' image is removed.

Change-Id: I847aec3d6ead64c7be81e6a5e758cbe362fde77d