Move spec file back
This moves the spec file back into the source repository to allow
continuous testing of creating RPMs. This patch updates the spec file
with changes reflecting the next release.

Furthermore, the source line has been adjusted to avoid including the
release. The release is used for rpm packagers only and should not be
part of the source name.