Use Fedora copr to automatically create builds
This patch removes a clutch to build RPMs for just testing purposes in
Travis to know if the build is potentially broken. The configuration
used was only Fedora 24 and failed.

Rethinking the whole situation, this patch removes the clutch to only
leave executing the tests in travis, but puts the RPM building in the
responsibility of Fedora copr.