Update manpages test for removal of config file check
This fixes the test suite for my previous commit, removing the
requirement for config files to have a manpage. One test was
specifically testing that check, so it is removed. Two other
tests relied on the behaviour, so I changed them to use an
executable under /bin instead of a config file.

Signed-off-by: Adam Williamson <awilliam@redhat.com>