Resync spec file with Fedora package repo, fix %check
The spec files in this repo and the downstream Fedora package
repo are meant, I think, to be in sync, but they've got a bit
out of sync over time. Only this repo has the %check section, so
we need to add that downstream; there also some changes in
a135cbf that were never sent downstream, so we'll need to do
those too.

On the other hand, c4ced6b mistakenly removed the dependency on
clamav itself; I corrected that in the downstream spec, so we
need to sync that change here.

I also noticed that %check appears to have been added in the
wrong place, slightly - it's right in the middle of %install,
not after it. The `packlist` and `fixperms` lines are supposed
to be part of %install, not %check, but the way %check was added
they wound up in %check. So, this fixes that too.

This version should hopefully be the best of all worlds!

Signed-off-by: Adam Williamson <awilliam@redhat.com>